refactor vector search tests to share mock dataset
