+++ tests/helpers/vectorSearch.test.js
@@ -1,32 +1,17 @@
 // @vitest-environment node
 import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
 import { withAllowedConsole } from "../utils/console.js";
+import { setupMockDataset } from "./vectorSearch/mockDataset.js";
 
 vi.mock("../../src/helpers/dataUtils.js", () => ({
   fetchJson: vi.fn(),
   validateWithSchema: vi.fn().mockResolvedValue(undefined)
 }));
 
-const sample = [
-  { id: "a", text: "A", embedding: [1, 0], source: "doc1", tags: ["foo"] },
-  { id: "b", text: "B", embedding: [0, 1], source: "doc2", tags: ["bar"] }
-];
-
 let fetchJsonMock;
 
 beforeEach(async () => {
-  vi.resetModules();
-  fetchJsonMock = (await import("../../src/helpers/dataUtils.js")).fetchJson;
-  fetchJsonMock.mockReset();
-  fetchJsonMock.mockImplementation((url) => {
-    if (url.endsWith("client_embeddings.manifest.json")) {
-      return Promise.resolve({ shards: ["shard1.json"] });
-    }
-    if (url.endsWith("shard1.json")) {
-      return Promise.resolve(sample);
-    }
-    return Promise.resolve([]);
-  });
+  fetchJsonMock = await setupMockDataset();
 });
 
 afterEach(() => {
@@ -80,5 +65,20 @@ describe("vectorSearch scoring", () => {
     const { findMatches } = await import("../../src/helpers/vectorSearch/scorer.js");
     const res = await findMatches([1, 0, 0], 1);
     expect(res).toEqual([]);
+  });
+
+  it("ensures embeddings use three-decimal precision", async () => {
+    const decimalSample = [
+      { id: "c", text: "C", embedding: [0.123, 0.456, 0.789], source: "doc3" }
+    ];
+    fetchJsonMock = await setupMockDataset(decimalSample);
+    const { loadEmbeddings } = await import("../../src/helpers/vectorSearch/loader.js");
+    const embeddings = await loadEmbeddings();
+    for (const entry of embeddings ?? []) {
+      for (const value of entry.embedding) {
+        const decimals = value.toString().split(".")[1];
+        expect(decimals ? decimals.length : 0).toBeLessThanOrEqual(3);
+      }
+    }
   });
 });
